requirements.txt
asgiref==3.5.2
asttokens==2.0.5
backcall==0.2.0
backports.zoneinfo==0.2.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.1
Django==4.0.4
djangorestframework==3.13.1
drf-yasg==1.20.0
executing==0.8.3
idna==3.3
importlib-metadata==4.11.4
inflection==0.5.1
ipdb==0.13.9
ipython==8.4.0
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.2
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
openapi-codec==1.3.2
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.29
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.12.0
pyparsing==3.0.9
pytz==2022.1
PyYAML==6.0
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
toml==0.10.2
traitlets==5.2.2.post1
uritemplate==4.1.1
urllib3==1.26.9
wcwidth==0.2.5
zipp==3.8.0
